refactor: Split site logic out
